requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coverage==7.13.2
cryptography==46.0.3
defusedxml==0.7.1
dnspython==2.8.0
eventlet==0.36.1
Flask==3.1.2
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
fonttools==4.60.2
fpdf2==2.8.4
greenlet==3.3.1
h11==0.16.0
idna==3.11
importlib_metadata==8.7.1
influxdb-client==1.43.0
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==11.3.0
pluggy==1.6.0
ply==3.11
psycopg2-binary==2.9.9
pyasn1==0.6.2
pycparser==3.0
Pygments==2.19.2
pymodbus==3.6.8
pysmi-lextudio==1.4.3
pysnmp==7.1.22
pytest==9.0.2
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
python-engineio==4.13.0
python-socketio==5.16.0
reactivex==4.1.0
reportlab
requests==2.32.5
setuptools==80.10.2
simple-websocket==1.1.0
six==1.17.0
t61codec==1.0.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
wsproto==1.3.2
x690==1.0.0.post1
zipp==3.23.0
bcrypt==4.2.0
